Refactor image mounter and location simulation
